Migrate git function tests
